# src/users/dto/change_password.dto.ts
import { ApiModelProperty } from "@nestjs/swagger";

export class ChangePasswordDto {
  
  @ApiModelProperty()
  readonly old_password: string;

  @ApiModelProperty()
  readonly new_password: string;

  

}