# swagger-spec.json
{"swagger":"2.0","info":{"description":"Node API Template","version":"1.0","title":"Node API Template"},"basePath":"/","tags":[],"schemes":["http","https"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/":{"get":{"responses":{"200":{"description":""}},"produces":["application/json"],"consumes":["application/json"]}},"/swagger-spec.json":{"get":{"responses":{"200":{"description":""}},"produces":["application/json"],"consumes":["application/json"]}},"/users/register":{"post":{"parameters":[{"name":"CreateUserDto","required":true,"in":"body","schema":{"$ref":"#/definitions/CreateUserDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/users/login":{"post":{"parameters":[{"name":"LoginUserDto","required":true,"in":"body","schema":{"$ref":"#/definitions/LoginUserDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/users/logOut":{"get":{"responses":{"200":{"description":""}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/users/view":{"get":{"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/users/update":{"post":{"parameters":[{"name":"UpdateUserDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateUserDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/users/change_password":{"post":{"parameters":[{"name":"ChangePasswordDto","required":true,"in":"body","schema":{"$ref":"#/definitions/ChangePasswordDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/users/forgetPassword":{"post":{"parameters":[{"name":"ForgetPasswordDto","required":true,"in":"body","schema":{"$ref":"#/definitions/ForgetPasswordDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/users/resetPassword":{"post":{"parameters":[{"name":"ForgetPasswordResetDto","required":true,"in":"body","schema":{"$ref":"#/definitions/ForgetPasswordResetDto"}}],"responses":{"201":{"description":""}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/gii":{"get":{"responses":{"200":{"description":""}},"tags":["gii"],"produces":["application/json"],"consumes":["application/json"]},"post":{"parameters":[{"name":"checkCollectionNameDto","required":true,"in":"body","schema":{"$ref":"#/definitions/checkCollectionNameDto"}}],"responses":{"201":{"description":""}},"tags":["gii"],"produces":["application/json"],"consumes":["application/json"]}},"/gii/createFiles":{"post":{"parameters":[{"name":"checkCollectionNameDto","required":true,"in":"body","schema":{"$ref":"#/definitions/checkCollectionNameDto"}}],"responses":{"201":{"description":""}},"tags":["gii"],"produces":["application/json"],"consumes":["application/json"]}},"/collection/create":{"post":{"parameters":[{"name":"CreateDto","required":true,"in":"body","schema":{"$ref":"#/definitions/CreateDto"}}],"responses":{"201":{"description":""}},"tags":["collection"],"produces":["application/json"],"consumes":["application/json"]}},"/collection/update":{"post":{"parameters":[{"name":"UpdateDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateDto"}}],"responses":{"201":{"description":""}},"tags":["collection"],"produces":["application/json"],"consumes":["application/json"]}},"/collection/view/{id}":{"get":{"parameters":[{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["collection"],"produces":["application/json"],"consumes":["application/json"]}},"/collection/list/{skip}/{limit}":{"get":{"parameters":[{"type":"number","name":"skip","required":true,"in":"path"},{"type":"number","name":"limit","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["collection"],"produces":["application/json"],"consumes":["application/json"]}},"/collection/delete":{"delete":{"parameters":[{"name":"DeleteDto","required":true,"in":"body","schema":{"$ref":"#/definitions/DeleteDto"}}],"responses":{"200":{"description":""}},"tags":["collection"],"produces":["application/json"],"consumes":["application/json"]}},"/sample/create":{"post":{"parameters":[{"name":"CreateDto","required":true,"in":"body","schema":{"$ref":"#/definitions/CreateDto"}}],"responses":{"201":{"description":""}},"tags":["sample"],"produces":["application/json"],"consumes":["application/json"]}},"/sample/update":{"post":{"parameters":[{"name":"UpdateDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateDto"}}],"responses":{"201":{"description":""}},"tags":["sample"],"produces":["application/json"],"consumes":["application/json"]}},"/sample/view/{id}":{"get":{"parameters":[{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["sample"],"produces":["application/json"],"consumes":["application/json"]}},"/sample/list/{skip}/{limit}":{"get":{"parameters":[{"type":"number","name":"skip","required":true,"in":"path"},{"type":"number","name":"limit","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["sample"],"produces":["application/json"],"consumes":["application/json"]}},"/sample/delete":{"delete":{"parameters":[{"name":"DeleteDto","required":true,"in":"body","schema":{"$ref":"#/definitions/DeleteDto"}}],"responses":{"200":{"description":""}},"tags":["sample"],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"Date":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","example":"user || admin || super_admin"},"created_date_time":{"title":"Date","allOf":[{"$ref":"#/definitions/Date"},{"example":"2019-05-23T10:02:23.059Z"}]}},"required":["first_name","last_name","email","password","role","created_date_time"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"}},"required":["first_name","last_name","email"]},"ChangePasswordDto":{"type":"object","properties":{"old_password":{"type":"string"},"new_password":{"type":"string"}},"required":["old_password","new_password"]},"ForgetPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ForgetPasswordResetDto":{"type":"object","properties":{"password":{"type":"string"},"token":{"type":"string"}},"required":["password","token"]},"checkCollectionNameDto":{"type":"object","properties":{"name":{"type":"string"},"field_name":{"type":"array","items":{"type":"string"}},"field_type":{"type":"array","items":{"type":"string"}},"field_required":{"type":"array","items":{"type":"string"}}},"required":["name","field_name","field_type","field_required"]},"CreateDto":{"type":"object","properties":{"sample2":{"type":"number"},"sample1":{"type":"string"}},"required":["sample2","sample1"]},"UpdateDto":{"type":"object","properties":{"id":{"type":"string"},"sample2":{"type":"number"},"sample1":{"type":"string"}},"required":["id","sample2","sample1"]},"DeleteDto":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]}}}